chore: rename `core` files to match class name
* chore: rename `core` files to match class name

* chore: change file import to match the renamed files
